Added cilium network policy to allow non-pod entities on remote nodes to access the cluster dns server.
In case cilium's dns based policy rules are utilized the dns traffic is redirected through a proxy component
owned by cilium. However, this proxy component needs to be allowed to perform dns requests. The default
network policy does not seem to be sufficient. Therefore, we add this special policy in addition.