Ensure that the kubernetes api server host is always set when running without kube-proxy.
During creation of a cluster there are a few races ongoing. This is especially true if
cilium is running as kube-proxy replacement.
cilium-agent waits for the cilium CRDs to be present in the kubernetes api server, before
performing its ordinary operation, i.e. there is no service routing in case kube-proxy is
disabled and no CNI operations will succeed.
cilium-operator is responsible for adding the CRDs on startup. However, unless the kubernetes
service host is specified, it will fall back on the in cluster name of the kubernetes api
server, i.e. kubernetes.default.svc.cluster.local. Unfortunately, this might not be
resolvable as cilium-agent might still wait on cilium-operator to create the CRDs. This
can lead to a vicious cycle and a hanging cluster creation.
Another contender, which might break the deadlock in some situations, can be the
apiserver-proxy-webhook, which can add the environment variable KUBERNETES_SERVICE_HOST
to containers. However, it is not ensured and actually not too likely that it is
ready before the cilium-operator pod is created.
To resolve this potential deadlock, this change ensures that the kubernetes api server
host is set for the cilium-operator when kube-proxy is disabled. The api server address
is taken from the shoot status, where the external address is utilized. Specific settings
in the network config can override this default.