Fix mutating web hook for node-local-dns to work with non-default ports.
In case a non-default port, i.e. not port 8080, was used the specification
of the liveness probe and the exposed health check port may not fit after
the mutating web hook for node-local-dns has been applied. This should be
fixed now for the standard case.
(Port 8080 was not a good choice for a host port and therefore it was
changed.)