Use resources.gardener.cloud/delete-on-invalid-update annotation for hubble generate certificate job. (#144)
Changes in the generate certificate job caused issues as the spec.template of the job is immutable.
The gardener resource manager allows to workaround this limitation when a resource is annotated with
resources.gardener.cloud/delete-on-invalid-update=true. By setting this annotation, it is possible
to properly use a container image overwrite or update the resource on upstream updates.