Fix reconciliation of hibernated clusters.
In the past, the extension library utilized a lazy kubernetes client instantiation.
However, this changed recently with gardener v1.80 and its controller-runtime update.
Now, the kubernetes client creation will fail if the api-server is not available.
In case a shoot cluster is hibernated, the api-server is not available. Hence, the
reconciliation will run into a corresponding error. This was handled previously by
discarding the error returned from the GET call. Now, we check first if the cluster
is hibernated and only create the client if it is not.