Clean-up iptables nat rules.
IPv4NativeRoutingCIDR = "0.0.0.0/0" leads to an invalid iptables rule that never matches. In case of iptables-nft this leads to an error.
The intention of choosing "0.0.0.0/0" was to have a CIDR that includes both pod and nodes range.
However, cilium takes care not to masquerade traffic to node addresses.
So it's save to choose the pod range for IPv4NativeRoutingCIDR.
As a result, our own iptables masquerade rules are not needed anymore. SnatOutOfCluster and SnatToUpstreamDNS can be switched off as default.