Allow usage of extension in Autonomous Shoot Cluster scenario.
Autonomous shoot clusters do not have the equivalent of the gardener
secret containing a kubeconfig for the control plane. As seed and shoot
cluster are basically the same thing in this scenario there is no need
for this secret. Therefore, extensions should work with what they have.
While it is possible to just use the seed client instead, in this case
it is also possible to simply not retrieve the config maps at all
because there is anyway a fallback logic in place. This way the code is
not cluttered with autonomous shoot cluster specifics, but will continue
to work nevertheless.