Fix a bug that prevented prometheus from storing cilium metrics. (#608)
In 6c7ddad, an additional snippet was added to the metricRelabelings part
of the Prometheus scrape config that instructed Prometheus to keep only
metrics where the `namespace` label was set to `kube-system`. When
Prometheus is configured to `keep` a subset of the metrics, it drops all
metrics that aren't kept.
The issue with this is that the `__meta_kubernetes_namespace` label isn't
replaced with the `namespace` label. In fact that would be unnecessary
because the calico-node pods only run in the `kube-system` namespace.
Therefore this commit simply removes the `keep` instruction for the
non-existent `namespace` label.